Relocate container_traits/_ohmms/_multi.h
