+++ src/AFQMC/HamiltonianOperations/THCOpsIO.hpp
@@ -18,7 +18,6 @@
 
 #include <fstream>
 
-#include "type_traits/container_traits_multi.h"
 #include "hdf/hdf_multi.h"
 #include "hdf/hdf_archive.h"
