+++ src/AFQMC/Hamiltonians/THCHamiltonian.cpp
@@ -9,7 +9,6 @@
 #include <numeric>
 
 #include "Configuration.h"
-#include "type_traits/container_traits_multi.h"
 #include "hdf/hdf_multi.h"
 #include "hdf/hdf_archive.h"
