+++ src/AFQMC/Matrix/ma_hdf5_readers.hpp
@@ -26,7 +26,6 @@
 #endif
 
 #include "Utilities/FairDivide.h"
-#include "type_traits/container_traits_multi.h"
 #include "hdf/hdf_multi.h"
 #include "hdf/hdf_archive.h"
