+++ src/io/hdf/hdf_hyperslab.h
@@ -14,7 +14,6 @@
 #define QMCPLUSPLUS_HDF_HYPERSLAB_IO_H
 
 #include <array>
-#include "type_traits/container_traits.h"
 #include "hdf_datatype.h"
 #include "hdf_dataspace.h"
 #include "hdf_dataproxy.h"
