+++ src/io/hdf/hdf_pete.h
@@ -13,10 +13,11 @@
 #ifndef QMCPLUSPLUS_HDF_PETE_TRAITS_H
 #define QMCPLUSPLUS_HDF_PETE_TRAITS_H
 
-#include "hdf_dataproxy.h"
+#include "container_traits_ohmms.h"
 #include "OhmmsPETE/OhmmsVector.h"
 #include "OhmmsPETE/OhmmsMatrix.h"
 #include "OhmmsPETE/OhmmsArray.h"
+#include "hdf_dataproxy.h"
 
 namespace qmcplusplus
 {
