+++ src/io/hdf/tests/test_hdf_archive.cpp
@@ -17,7 +17,6 @@
 #include <vector>
 #include <string>
 #include <sstream>
-#include "type_traits/container_traits_ohmms.h"
 
 using std::string;
 using std::vector;
